3d_suite/requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
plotly>=5.0.0
fpdf2>=2.5.0
openpyxl>=3.0.0
requests-oauthlib>=1.3.0
pydantic>=1.8.0